clean up nivo config
